Updating UserPolicy message
The UserPolicy message can either use a static authorized user list by
account, or can be configured to use an AuthorizedPrincipalsCommand.
Rather than a hard coded file path and arguments, we select the dynamic
checker using an enum and a vendor would need to determine out to
satisfy this strategy. HIBA is the first and only such strategy so far.