certz: add CSRSuite enums
The possible combinations of KeyType, SignatureAlgorithm
and CertificateType (and key_size_bits for RSA)
have been consolidated into enums.
This simplifies the request for the client, by
hardcoding the types of CSR requests that can be sent.

key_size_bits also had ambiguity for non-RSA keys,
which is resolved by including information about the
key size in the enum name.

The enum names also specify that SHA-2 is being used,
so as to be forward compatible if SHA-3 is adopted
in future

This commit also updates the certz.pb.go file to
 include these changes