acctz: update GrpcService payload field
Udpate the payload field to be a string value.
This is used instead of google.protobuf.Any as
google.protobuf.Any cannot be unmarshalled into
a known proto message type if it is truncated
(as google.protobuf.Any is encoded).

Also add a comment specifying that this value
may be partially redacted in the case where we
want to avoid leaking sensitive information from
the device (such as private keys rotated via
gNSI Certz or gNSI Credentialz).