certz: clarify usage of entity_type field
clarify that we should return an error if the
entity_type field matches the type of an entity already
being rotated in the same UploadRequest.
I.e. the artifact being rotated should either be uploaded
in the UploadRequest _or_ it should be copied from
an existing ssl profile, but not both.